Substituindo classes Template por Entities
#Resolve #19

+++ Gameserver/src/main/com.l2jbr.gameserver/com/l2jbr/gameserver/model/actor/instance/L2NpcInstance.java
@@ -2175,11 +2175,11 @@ public class L2NpcInstance extends L2Character {
         _currentRHandId = newWeaponId;
     }
 
-    public void setCollisionHeight(int height) {
+    public void setCollisionHeight(float height) {
         _currentCollisionHeight = height;
     }
 
-    public void setCollisionRadius(int radius) {
+    public void setCollisionRadius(float radius) {
         _currentCollisionRadius = radius;
     }
